Move the `Checkable` typeclass into a new `anticipation.check`
`quantitative` depended on `probably` — the test framework — for one reason:
it defines a `Checkable` instance for quantities, and builds `Checkable`
values in a macro. That put the whole test framework, and with it
chiaroscuro, escritoire, eucalyptus and dendrology, in the closure of
abacist, aviation, mosquito, phoenicia and everything above them.

`Checkable`, `Tolerance` and the `===`, `!==`, `+/-` and `±` operators move to
a new `anticipation.check`, which is what an anticipation module is for: the
interface a module needs in order to avoid depending on an implementation.
`probably.core` now depends on it and re-exports the names, so `import
probably.*` is unchanged for every existing call site, and the givens stay in
`Checkable`'s companion, so implicit scope is untouched.

phoenicia and facsimile were reaching turbulence and eucalyptus through
quantitative's dependency on probably, and use both directly, so they now
declare them. xylophone's tests decode an `EmailAddress`, so they take the
`urticose.core` dependency that core no longer has.
