+++ lib/anticipation/src/check/anticipation.Tolerance.scala
@@ -30,7 +30,7 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package probably
+package anticipation
 
 case class Tolerance[value](base: value, tolerance: value)
   ( greaterThan: (value, value) => Boolean,
