+++ lib/anticipation/src/check/anticipation_check.scala
@@ -30,18 +30,41 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package soundness
+package anticipation
 
-export
-  probably
-  . { Anchor, Arithmetic, Autopsy, Axable, Axis, Baseline, Benchmark, Cadential,
-      Ci, debug, Geometric, GithubActions, Harness, Hotspots, Inclusion,
-      Max, Mean, Metric, Min, nominative, Probing, Report, Reporter, Run, Runner, Spread,
-      Spread2, Strain, suite, Tally, Temporal, Test, test, Testable, TestId, TestPalette,
-      Trial, Value, Verdict }
+import scala.annotation.*
 
-package harnesses:
-  export probably.harnesses.threadLocal
+import hypotenuse.*
+import prepositional.*
+import symbolism.*
 
-package autopsies:
-  export probably.autopsies.{contrastExpectations, none}
+extension [left](left: left)
+  infix def === [right](right: right)(using checkable: left is Checkable against right): Boolean =
+    checkable.check(left, right)
+
+  infix def !== [right](right: right)(using checkable: left is Checkable against right): Boolean =
+    !checkable.check(left, right)
+
+extension [value](value: value)
+  @targetName("plusOrMinus")
+  inline infix def +/- (tolerance: value)
+  ( using inline commensurable: value is Commensurable against value,
+          addable:              value is Addable by value,
+          equality:             addable.Result =:= value,
+          subtractable:         value is Subtractable by value,
+          equality2:            subtractable.Result =:= value )
+  :   Tolerance[value] =
+
+    Tolerance[value](value, tolerance)(_ >= _, _ + _, _ - _)
+
+
+  @targetName("plusOrMinus2")
+  inline infix def ± (tolerance: value)
+    ( using inline commensurable: value is Commensurable against value,
+            addable:              value is Addable by value,
+            equality:             addable.Result =:= value,
+            subtractable:         value is Subtractable by value,
+            equality2:            subtractable.Result =:= value )
+  :   Tolerance[value] =
+
+    value +/- (tolerance)
