+++ lib/quantitative/src/core/quantitative.protointernal.scala
@@ -45,7 +45,6 @@ import fulminate.*
 import gigantism.*
 import hypotenuse.*
 import prepositional.*
-import probably.*
 import rudiments.*
 import symbolism.*
